backend-py-main/requirements.txt
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.9.0
awscli==1.37.26
bcrypt==4.3.0
blinker==1.9.0
botocore==1.36.26
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.2
dnspython==2.7.0
docutils==0.16
ecdsa==0.19.1
et_xmlfile==2.0.0
fastapi==0.115.11
Flask==3.1.0
flask-cors==5.0.1
Flask-MySQLdb==2.0.0
Flask-PyMongo==3.0.1
Flask-SQLAlchemy==3.1.1
google-api-core==2.24.2
google-auth==2.38.0
google-cloud-speech==2.31.1
googleapis-common-protos==1.69.2
greenlet==3.1.1
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.14.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
jmespath==1.0.1
MarkupSafe==3.0.2
motor==3.7.0
mysqlclient==2.2.7
numpy==2.2.4
openpyxl==3.1.5
pandas==2.2.3
passlib==1.7.4
proto-plus==1.26.1
protobuf==5.29.4
pyasn1==0.4.8
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.10.1
pymongo==4.11.3
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
requests==2.32.3
rsa==4.7.2
s3transfer==0.11.2
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.38
starlette==0.46.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
Werkzeug==3.1.3
XlsxWriter==3.2.2
